requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
gspread>=5.11.0
openpyxl>=3.1.2
bcrypt>=4.0.0
PyYAML>=6.0
python-dateutil
pymupdf>=1.23.0
reportlab>=4.0.0
Pillow>=10.0.0
streamlit-drawable-canvas>=0.9.3
requests>=2.31.0